Exercicio 3 completo

+++ Form3.vb
@@ -2,7 +2,7 @@
     Private Sub btnCalcularVolume_Click(sender As Object, e As EventArgs) Handles btnCalcularVolume.Click
         Dim volume As Double
 
-        volume = Convert.ToInt32(txtBoxAltura.Text) * Convert.ToInt32(txtBoxComprimento.Text) * Convert.ToInt32(txtBoxLargura.Text)
+        volume = Convert.ToDouble(txtBoxAltura.Text) * Convert.ToDouble(txtBoxComprimento.Text) * Convert.ToDouble(txtBoxLargura.Text)
 
         txtBoxResultadoVolume.Text = volume
